Khoa update 23/09

+++ src/main/java/com/tdkhoa/ecommerce/services/impl/ReviewServiceImpl.java
@@ -4,11 +4,19 @@
  */
 package com.tdkhoa.ecommerce.services.impl;
 
+import com.cloudinary.Cloudinary;
+import com.cloudinary.utils.ObjectUtils;
+import com.tdkhoa.ecommerce.Pojo.Product;
 import com.tdkhoa.ecommerce.Pojo.Review;
+import com.tdkhoa.ecommerce.Pojo.User;
+import com.tdkhoa.ecommerce.repositories.ProductRepository;
 import com.tdkhoa.ecommerce.repositories.ReviewRepository;
 import com.tdkhoa.ecommerce.services.ReviewService;
+import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,48 +28,65 @@ import org.springframework.web.multipart.MultipartFile;
  */
 @Service
 public class ReviewServiceImpl implements ReviewService {
+
     @Autowired
     private ReviewRepository rRepo;
-    
+    @Autowired
+    private Cloudinary cloudinary;
+    @Autowired
+    private ProductRepository pRepo;
+
     @Override
-    public List<Review> getListsReview() {
+    public List<Review> getListsReviewByProductId(int idProduct) {
         List<Review> reviews = this.rRepo.findAll();
         return reviews;
     }
-    
+
     @Override
-    public boolean add(Map<String, String> params, MultipartFile imageUrl) {
+    public boolean add(Map<String, String> params, MultipartFile imageUrl, User user, int idProduct) {
         Review r = new Review();
-//        if (!imageUrl.isEmpty()) {
-//            try {
-//                Map res = this.cloudinary.uploader().upload(imageUrl.getBytes(), ObjectUtils.asMap("resource_type", "auto"));
-//                p.setImageUrl(res.get("secure_url").toString());
-//            } catch (IOException ex) {
-//                Logger.getLogger(BannerService.class.getName()).log(Level.SEVERE, null, ex);
-//            }
-//        }
+        r.setContent(params.get("content"));
+        r.setUser(user);
+        Product p = this.pRepo.findById(idProduct).get();
+        r.setProduct(p);
+        if (!imageUrl.isEmpty()) {
+            try {
+                Map res = this.cloudinary.uploader().upload(imageUrl.getBytes(), ObjectUtils.asMap("resource_type", "auto"));
+                r.setImageUrl(res.get("secure_url").toString());
+            } catch (IOException ex) {
+                Logger.getLogger(ReviewServiceImpl.class.getName()).log(Level.SEVERE, null, ex);
+            }
+        }
         this.rRepo.save(r);
         return true;
     }
-    
+
     @Override
-    public boolean update(Map<String, String> params, MultipartFile imageUrl, @PathVariable(value = "id") int id) {
+    public boolean update(Map<String, String> params, MultipartFile imageUrl, @PathVariable(value = "id") int id, User user) {
         Review r = this.rRepo.findById(id).get();
-//        if (!imageUrl.isEmpty()) {
-//            try {
-//                Map res = this.cloudinary.uploader().upload(imageUrl.getBytes(), ObjectUtils.asMap("resource_type", "auto"));
-//                p.setImageUrl(res.get("secure_url").toString());
-//            } catch (IOException ex) {
-//                Logger.getLogger(BannerService.class.getName()).log(Level.SEVERE, null, ex);
-//            }
-//        }
-        this.rRepo.save(r);
-        return true;
+        if (r.getUser().equals(user)) {
+            r.setContent(params.get("content"));
+            if (!imageUrl.isEmpty()) {
+                try {
+                    Map res = this.cloudinary.uploader().upload(imageUrl.getBytes(), ObjectUtils.asMap("resource_type", "auto"));
+                    r.setImageUrl(res.get("secure_url").toString());
+                } catch (IOException ex) {
+                    Logger.getLogger(ReviewServiceImpl.class.getName()).log(Level.SEVERE, null, ex);
+                }
+            }
+            this.rRepo.save(r);
+            return true;
+        }
+        return false;
     }
-    
+
     @Override
-    public boolean delete(int id) {
-        this.rRepo.delete(this.rRepo.findById(id).get());
-        return true;
+    public boolean delete(int id, User user) {
+        Review r = this.rRepo.findById(id).get();
+        if (r.getUser().equals(user)) {
+            this.rRepo.delete(this.rRepo.findById(id).get());
+            return true;
+        }
+        return false;
     }
 }
